Update: Move receivers and test structs away to prepare new test cases for RFC6587
