+++ plugins/inputs/syslog/rfc5425_test.go
@@ -16,20 +16,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-var (
-	pki = testutil.NewPKI("../../../testutil/pki")
-)
-
-type testCase5425 struct {
-	name           string
-	data           []byte
-	wantBestEffort []testutil.Metric
-	wantStrict     []testutil.Metric
-	werr           int // how many errors we expect in the strict mode?
-}
-
-func getTestCasesForRFC5425() []testCase5425 {
-	testCases := []testCase5425{
+func getTestCasesForRFC5425() []testCaseStream {
+	testCases := []testCaseStream{
 		{
 			name: "1st/avg/ok",
 			data: []byte(`188 <29>1 2016-02-21T04:32:57+00:00 web1 someservice 2341 2 [origin][meta sequence="14125553" service="someservice"] "GET /v1/ok HTTP/1.1" 200 145 "-" "hacheck 0.9.0" 24306 127.0.0.1:40124 575`),
@@ -346,35 +334,11 @@ func getTestCasesForRFC5425() []testCase5425 {
 	return testCases
 }
 
-func newTCPSyslogReceiver(address string, keepAlive *internal.Duration, maxConn int, bestEffort bool) *Syslog {
-	d := &internal.Duration{
-		Duration: defaultReadTimeout,
-	}
-	s := &Syslog{
-		Address: address,
-		now: func() time.Time {
-			return defaultTime
-		},
-		TransparentFraming: true,
-		ReadTimeout:        d,
-		BestEffort:         bestEffort,
-		Separator:          "_",
-	}
-	if keepAlive != nil {
-		s.KeepAlivePeriod = keepAlive
-	}
-	if maxConn > 0 {
-		s.MaxConnections = maxConn
-	}
-
-	return s
-}
-
 func testStrictRFC5425(t *testing.T, protocol string, address string, wantTLS bool, keepAlive *internal.Duration) {
 	for _, tc := range getTestCasesForRFC5425() {
 		t.Run(tc.name, func(t *testing.T) {
 			// Creation of a strict mode receiver
-			receiver := newTCPSyslogReceiver(protocol+"://"+address, keepAlive, 0, false)
+			receiver := newTCPSyslogReceiver(protocol+"://"+address, keepAlive, 0, false, true)
 			require.NotNil(t, receiver)
 			if wantTLS {
 				receiver.ServerConfig = *pki.TLSServerConfig()
@@ -436,7 +400,7 @@ func testBestEffortRFC5425(t *testing.T, protocol string, address string, wantTL
 	for _, tc := range getTestCasesForRFC5425() {
 		t.Run(tc.name, func(t *testing.T) {
 			// Creation of a best effort mode receiver
-			receiver := newTCPSyslogReceiver(protocol+"://"+address, keepAlive, 0, true)
+			receiver := newTCPSyslogReceiver(protocol+"://"+address, keepAlive, 0, true, true)
 			require.NotNil(t, receiver)
 			if wantTLS {
 				receiver.ServerConfig = *pki.TLSServerConfig()
@@ -487,31 +451,31 @@ func testBestEffortRFC5425(t *testing.T, protocol string, address string, wantTL
 	}
 }
 
-func TestStrict_tcp(t *testing.T) {
+func TestRFC5425Strict_tcp(t *testing.T) {
 	testStrictRFC5425(t, "tcp", address, false, nil)
 }
 
-func TestBestEffort_tcp(t *testing.T) {
+func TestRFC5425BestEffort_tcp(t *testing.T) {
 	testBestEffortRFC5425(t, "tcp", address, false, nil)
 }
 
-func TestStrict_tcp_tls(t *testing.T) {
+func TestRFC5425Strict_tcp_tls(t *testing.T) {
 	testStrictRFC5425(t, "tcp", address, true, nil)
 }
 
-func TestBestEffort_tcp_tls(t *testing.T) {
+func TestRFC5425BestEffort_tcp_tls(t *testing.T) {
 	testBestEffortRFC5425(t, "tcp", address, true, nil)
 }
 
-func TestStrictWithKeepAlive_tcp_tls(t *testing.T) {
+func TestRFC5425StrictWithKeepAlive_tcp_tls(t *testing.T) {
 	testStrictRFC5425(t, "tcp", address, true, &internal.Duration{Duration: time.Minute})
 }
 
-func TestStrictWithZeroKeepAlive_tcp_tls(t *testing.T) {
+func TestRFC5425StrictWithZeroKeepAlive_tcp_tls(t *testing.T) {
 	testStrictRFC5425(t, "tcp", address, true, &internal.Duration{Duration: 0})
 }
 
-func TestStrict_unix(t *testing.T) {
+func TestRFC5425Strict_unix(t *testing.T) {
 	tmpdir, err := ioutil.TempDir("", "telegraf")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmpdir)
@@ -519,7 +483,7 @@ func TestStrict_unix(t *testing.T) {
 	testStrictRFC5425(t, "unix", sock, false, nil)
 }
 
-func TestBestEffort_unix(t *testing.T) {
+func TestRFC5425BestEffort_unix(t *testing.T) {
 	tmpdir, err := ioutil.TempDir("", "telegraf")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmpdir)
@@ -527,7 +491,7 @@ func TestBestEffort_unix(t *testing.T) {
 	testBestEffortRFC5425(t, "unix", sock, false, nil)
 }
 
-func TestStrict_unix_tls(t *testing.T) {
+func TestRFC5425Strict_unix_tls(t *testing.T) {
 	tmpdir, err := ioutil.TempDir("", "telegraf")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmpdir)
@@ -535,7 +499,7 @@ func TestStrict_unix_tls(t *testing.T) {
 	testStrictRFC5425(t, "unix", sock, true, nil)
 }
 
-func TestBestEffort_unix_tls(t *testing.T) {
+func TestRFC5425BestEffort_unix_tls(t *testing.T) {
 	tmpdir, err := ioutil.TempDir("", "telegraf")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmpdir)
